Create scrabble.c
this is the code that cs50 wrote and I only added some things to create a functioning program. I wrote a code that can give the appropriate points for the word that the user entered. The program will compare two words entered by two different players and calculate the points of each word based on the scrabble word points, it will then announce the player with the most points as the winner.