Create runoff.c
this is the hardest program ive ever encountered so far.. both of the problems this week 3 is so hard, and this is the easiest of the two. I even got 1 test wrong making me have 97% score. this program need you to count the votes for all the candidates depending on the ranking of the vote, print the tie if it happened, use the elimination method to know the winner. took me a 5hrs to figure out and still got 1 test wrong. I give up thinking on how to initiate multiple preferences when counting the votes when theres many eliminated candidates.