Create volume.c
this is my week 4 cs50 lab. it reads the 44 header bytes inside a WAV file and stores them in another WAV file. and then reads all the 2byte samples and multiply all of them based on the entered factor, and then write the changed sample bytes to the output WAV file. This will increase or decrease the volume of the WAV file